Use a continuation monad to cut out the use of CSE
