+++ vehicle/src/Vehicle/Compile/Queries.hs
@@ -18,8 +18,10 @@ import Vehicle.Compile.Prelude
 import Vehicle.Compile.QuantifierAnalysis (checkQuantifiersAndNegateIfNecessary)
 import Vehicle.Compile.Queries.DNF (convertToDNF)
 import Vehicle.Compile.Queries.IfElimination (eliminateIfs)
+import Vehicle.Compile.Queries.LinearSatisfactionProblem (generateCLSTProblem)
 import Vehicle.Compile.Queries.NetworkElimination
 import Vehicle.Compile.Queries.QuantifierLifting (liftAndRemoveQuantifiers)
+import Vehicle.Compile.Queries.Variable (UserVariable (..))
 import Vehicle.Compile.Queries.VariableReconstruction
 import Vehicle.Compile.Resource
 import Vehicle.Compile.Type (getPropertyInfo, getUnnormalised)
@@ -177,9 +179,24 @@ compileSingleQuery expr = do
     -- First lift all the quantifiers to the top-level and remove them.
     (quantLiftedExpr, userVariables) <- liftAndRemoveQuantifiers expr
 
-    -- Convert all user variables and applications of networks into magic I/O variables
-    clstQuery <- normUserVariables ident verifier networkCtx userVariables quantLiftedExpr
+    -- Substitute through the let-bound applications.
+    let boundCtx = fmap (Just . userVarName) userVariables
 
+    -- Convert all user variables and applications of networks into magic I/O variables
+    (metaNetwork, networkVariables, inputEqualities, queryExpr) <-
+      replaceNetworkApplications networkCtx boundCtx quantLiftedExpr
+
+    -- Convert it into a linear satisfaction problem in the user variables
+    let state =
+          ( networkCtx,
+            ident,
+            metaNetwork,
+            userVariables,
+            networkVariables
+          )
+    clstQuery <- generateCLSTProblem state inputEqualities queryExpr
+
+    -- Compile the query to the specific verifiers.
     flip traverseQuery clstQuery $ \(clstProblem, u, v) -> do
       queryDoc <- compileQuery verifier clstProblem
       logCompilerPassOutput queryDoc
